Implemented load instructions with indirect addressing

+++ src/cpu_reg.cpp
@@ -97,18 +97,18 @@ void CpuRegisters::set_hl(uint16_t val) {
 void CpuRegisters::set_sp(uint16_t val) {
 	SP = val;
 }
-void CpuRegisters::add_sp(uint8_t val) {
+void CpuRegisters::add_sp(uint16_t val) {
 	SP += val;
 }
-void CpuRegisters::dec_sp(uint8_t val) {
+void CpuRegisters::dec_sp(uint16_t val) {
 	SP -= val;
 }
 void CpuRegisters::set_pc(uint16_t val) {
 	PC = val;
 }
-void CpuRegisters::add_pc(uint8_t val) {
+void CpuRegisters::add_pc(uint16_t val) {
 	PC += val;
 }
-void CpuRegisters::dec_pc(uint8_t val) {
+void CpuRegisters::dec_pc(uint16_t val) {
 	PC -= val;
 }
